+++ src/silo/query_engine/operators/selection.test.cpp
@@ -6,11 +6,11 @@
 using silo::query_engine::operators::Selection;
 
 TEST(OperatorSelection, equalsShouldReturnCorrectValues) {
-   const std::vector<int64_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
+   const std::vector<int32_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
    const uint32_t row_count = 13;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::EQUALS, 1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::EQUALS, 1, row_count
    );
 
    ASSERT_EQ(*under_test.evaluate(), roaring::Roaring({1, 5, 6, 7, 8, 9}));
@@ -19,11 +19,11 @@ TEST(OperatorSelection, equalsShouldReturnCorrectValues) {
 }
 
 TEST(OperatorSelection, notEqualsShouldReturnCorrectValues) {
-   const std::vector<int64_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
+   const std::vector<int32_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
    const uint32_t row_count = 12;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::NOT_EQUALS, 1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::NOT_EQUALS, 1, row_count
    );
 
    ASSERT_EQ(*under_test.evaluate(), roaring::Roaring({0, 2, 3, 4}));
@@ -32,11 +32,11 @@ TEST(OperatorSelection, notEqualsShouldReturnCorrectValues) {
 }
 
 TEST(OperatorSelection, lessShouldReturnCorrectValues) {
-   const std::vector<int64_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
+   const std::vector<int32_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
    const uint32_t row_count = 15;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::LESS, 1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::LESS, 1, row_count
    );
 
    ASSERT_EQ(*under_test.evaluate(), roaring::Roaring({0}));
@@ -45,11 +45,11 @@ TEST(OperatorSelection, lessShouldReturnCorrectValues) {
 }
 
 TEST(OperatorSelection, lessOrEqualsShouldReturnCorrectValues) {
-   const std::vector<int64_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
+   const std::vector<int32_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
    const uint32_t row_count = 13;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::LESS_OR_EQUALS, 1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::LESS_OR_EQUALS, 1, row_count
    );
 
    ASSERT_EQ(*under_test.evaluate(), roaring::Roaring({0, 1, 5, 6, 7, 8, 9}));
@@ -58,11 +58,11 @@ TEST(OperatorSelection, lessOrEqualsShouldReturnCorrectValues) {
 }
 
 TEST(OperatorSelection, higherOrEqualsShouldReturnCorrectValues) {
-   const std::vector<int64_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
+   const std::vector<int32_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
    const uint32_t row_count = 13;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::HIGHER_OR_EQUALS, 1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::HIGHER_OR_EQUALS, 1, row_count
    );
 
    ASSERT_EQ(*under_test.evaluate(), roaring::Roaring({1, 2, 3, 4, 5, 6, 7, 8, 9}));
@@ -71,11 +71,11 @@ TEST(OperatorSelection, higherOrEqualsShouldReturnCorrectValues) {
 }
 
 TEST(OperatorSelection, higherShouldReturnCorrectValues) {
-   const std::vector<int64_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
+   const std::vector<int32_t> test_column({{0, 1, 4, 4, 4, 1, 1, 1, 1, 1}});
    const uint32_t row_count = 13;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::HIGHER, 1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::HIGHER, 1, row_count
    );
 
    ASSERT_EQ(*under_test.evaluate(), roaring::Roaring({2, 3, 4}));
@@ -84,22 +84,22 @@ TEST(OperatorSelection, higherShouldReturnCorrectValues) {
 }
 
 TEST(OperatorSelection, correctWithNegativeNumbers) {
-   const std::vector<int64_t> test_column({{0, -1, 4, 4, 4, -1, -1, -1, -1, -1}});
+   const std::vector<int32_t> test_column({{0, -1, 4, 4, 4, -1, -1, -1, -1, -1}});
    const uint32_t row_count = 13;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::EQUALS, -1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::EQUALS, -1, row_count
    );
 
    ASSERT_EQ(*under_test.evaluate(), roaring::Roaring({1, 5, 6, 7, 8, 9}));
 }
 
 TEST(OperatorSelection, returnsCorrectTypeInfo) {
-   const std::vector<int64_t> test_column({{0, -1, 4, 4, 4, -1, -1, -1, -1, -1}});
+   const std::vector<int32_t> test_column({{0, -1, 4, 4, 4, -1, -1, -1, -1, -1}});
    const uint32_t row_count = 13;
 
-   const Selection<int64_t> under_test(
-      test_column, Selection<int64_t>::Comparator::EQUALS, -1, row_count
+   const Selection<int32_t> under_test(
+      test_column, Selection<int32_t>::Comparator::EQUALS, -1, row_count
    );
 
    ASSERT_EQ(under_test.type(), silo::query_engine::operators::SELECTION);
